Send signaling message when starting or stopping breakout rooms
When the breakout room status changes a "room" message is sent to all
the active sessions in either the parent or the breakout rooms
(depending on whether they are being started or stopped) with the token
of the room that they have to switch to.

When the breakout rooms are started the message is sent only to non
moderators, as moderators do not have a single breakout room assigned.
On the other hand, when the breakout rooms are stopped the message is
also sent to all moderators (who are in a breakout room and not already
in the parent room), as all participants need to switch to the parent
room.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>