Handle "switchto" message in WebUI
When the client receives a message to switch to a different room the
WebUI joins that room. If the WebUI was already in a call it will
automatically join the call in the target room; in that case the call
view will be kept shown during the switch, rather than showing the chat
while leaving the previous call and joining the new room to then show
the call again when joining the call in the target room.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>